# test: benchmark the index merge performance