--- a/python/python/benchmarks/test_index.py
+++ b/python/python/benchmarks/test_index.py
@@ -69,11 +69,13 @@ def test_create_ivf_pq_cuda(test_dataset, benchmark):
 
 
 @pytest.mark.benchmark(group="optimize_index")
-@pytest.mark.parametrize("num_partitions", [32, 256, 512, 1024])
-@pytest.mark.parametrize("ef_construction", [150, 300, 500])
-@pytest.mark.parametrize("m", [16, 24])
-def test_optimize_index(test_large_dataset, benchmark, num_partitions, ef_construction, m):
-    # insert a smaller batch into the large dataset,
+@pytest.mark.parametrize("num_partitions", [512])
+@pytest.mark.parametrize("ef_construction", [150, 300])
+@pytest.mark.parametrize("m", [16])
+@pytest.mark.parametrize("num_small_indexes", [1, 5])
+@pytest.mark.parametrize("num_new_rows", [500])
+def test_optimize_index(test_large_dataset, benchmark, num_partitions, ef_construction, m, num_small_indexes, num_new_rows):
+    # insert smaller batch(es) into the large dataset,
     # then benchmark the optimize_index method
     m_max = m * 2
     test_large_dataset = test_large_dataset.create_index(
@@ -87,7 +89,9 @@ def test_optimize_index(test_large_dataset, benchmark, num_partitions, ef_constr
         num_bits=8,
         replace=True,
     )
-    small_table = gen_table(test_large_dataset.count_rows() * 30 // 100)
-    lance.write_dataset(small_table, test_large_dataset.uri, mode="append")
+
+    for _ in range(num_small_indexes):
+        small_table = gen_table(num_new_rows // num_small_indexes)
+        lance.write_dataset(small_table, test_large_dataset.uri, mode="append")
 
     benchmark(test_large_dataset.optimize.optimize_indices)