this is slow cause we now reconstruct the entire index after merging data.
for 512d 1M rows IVF_HNSW_SQ dataset:
merge 500 rows: 41.4195s
merge 5 batches (100 rows per batch): 42.3950s

for 512d 100K rows IVF_HNSW_SQ dataset:
merge 500 rows: 2.2307s
merge 5 batches (100 rows per batch): 2.2654s

for 512d 1M rows IVF_PQ dataset:
merge 500 rows: 174.4656ms
merge 5 batches (100 rows per batch): 178.6720ms